Changed method of activating preferred exchange.
Instead of using integer index offsets for storing the preferred exchange tab, each tab is now "tagged" with the exchange and stored as a preference item. This will allow future enhancements to allow specifying the visibility and tab re-ordering.

Also includes code to migrate from the old integer index method to the new tab tag method.